+++ src/Data/Status.php
@@ -13,12 +13,12 @@ final class Status extends Dto
 {
     public function __construct(
         public readonly string $name = '',
-        public readonly string $slug = '',
         public readonly bool $public = false,
         public readonly bool $protected = false,
         public readonly bool $private = false,
         public readonly bool $queryable = false,
         public readonly bool $show_in_list = false,
         public readonly bool $date_floating = false,
+        public readonly string $slug = '',
     ) {}
 }